[refactor] UI 파일 분할.
